+++ types/options.go
@@ -11,7 +11,7 @@ type DeployOptions struct {
 	Entrypoint string            // Entrypoint to deploy
 	ExtraArgs  string            // Extra arguments to append to command
 	CPUQuota   float64           // How many cores needed, e.g. 1.5
-	Count      int64             // How many containers needed, e.g. 4
+	Count      int               // How many containers needed, e.g. 4
 	Memory     int64             // Memory for container, in bytes
 	Env        []string          // Env for container
 	Networks   map[string]string // Network names and specified IPs
